requirements.txt
streamlit>=1.36.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
plyer>=2.1.0
pytz>=2024.1
